Functional reconstructor's accuracy validated
The implementation of `kesi.FunctionalKernelFieldReconstructor` is based
on the associativity of the matrix multiplication. The cross-kernel
matrix (known from the original kCSD implementation) is never
calculated explicitely. Instead, it is factorized: replaced by a product
of two factor matrices.

Both the right hand side factor matrix and the kernel matrix are
constant (they depend only on the location of the electrodes and the
current sources defined). Thus, given the recoreded potentials and
the regularization parameter, it is possible to calculate the rightmost
product of the CSD reconstruction just once for multiple sets of CSD
estimation points (which determine the left factor matrix only).

While such approach greatly improves performance and flexibility of
kESI, it has not been determined whether accuracy of both orders of
matrix multiplication are equivalent. The provided script validates,
whether the right-to-left is always more accurate in case of the actual
matrices used in examplary applications of kESI.