Leave-one-out cross validation implemented
`kesi.FunctionalKernelFieldReconstructor.leave_one_out_errors()` method
returns errors of leave-one-out crossvalidation.

Behaviour of `regularization_parameter` does not depend on the number of
field components ("sources"). If instances of exactly the the same
sources are multiplicated, the same value of parameter should lead to
the same results (differences due to the numerical errors are still
possible though).