`.leave_one_out_errors()` method improved
The method accepts time series of measurements (it is responsibility of
measurement manager to load them as a matrix which columns are vectors
of measurements at certain timepoints) and returns a sequence
of estimation errors (for the left out electrodes).  Estimation errors
are 1D `np.ndarrays` of errors for different timepoints.

If measurements are given as static (measurement manager loads them as
1D vector) they are interpreted as if they were 1 timepoint long time
series.  That behaviour may change.