Experimental Lanczos basis source provided
A factory class `fem_lanczos.LanczosSourceFactory` has been provided,
which loads FEM-simulated potentials for sources which CSD profile is
given as 3-dimensional product of Lanczos functions (in the direction of
orthogonal axes).

The module `fem_lanczos`, when run as program, calculates the potentials
and saves file to be loaded by the `LanczosSourceFactory` class.

Also, resolution of the used mesh has been improved.