`.potential_sinc_scalar()` method provided
The method mixes sinc interpolation with point source approximation and
is not vectorized to avoid sinc interpolation when unnecessary.

The approximation is mixed only for points further than
`R * approximation_mixing_threshold` where `R` is the maximal distance
for which the potential has been fully precalculated and
`approximation_mixing_threshold` is a real number from 0 to 1 given as
argument of `_fem_common._SymmetricSourceFactory_Base` subclass
constructor.