Provide degenerate point source factory
The DegeneratedSliceSourcesFactory has been provided to reduce memory
requirements of FEM-derived sources.

Warning: the code is not tested yet, as it has not been developed on the
production machine which is the only one able to test it. Git commit has
been used to synchronize code on both (production and development)
machines.